redis替换数据库

+++ src/main/java/com/xuanjia/millionlikes/model/domain/Blog.java
@@ -43,7 +43,7 @@ public class Blog {
     /**
      * 点赞数
      */
-    private Integer thumbCount;
+    private Long thumbCount;
 
     /**
      * 创建时间
